tests/wpt/css-tests/css-text-3_dev/xhtml1/white-space-collapse-000.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html lang="en" xmlns="http://www.w3.org/1999/xhtml"><head>
<meta charset="utf-8" />
<title>White space collapse</title>
<link href="mailto:ishida@w3.org" rel="author" title="Richard Ishida" />
<link href="https://drafts.csswg.org/css-text-3/#line-break-transform" rel="help" />
<meta content="Every tab is converted to a space. Any space immediately following another collapsible space is collapsed to have zero advance width." name="assert" />
<style type="text/css">
.test span { font-size: 24px; font-family: sans-serif; background-color: #2AA5F7; color: white; }
.ref span { font-size: 24px; font-family: sans-serif; background-color: #270CEF; color: white; }
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<div class="test" id="test"><span>hello   	       there</span></div>
<div class="ref" id="ref"><span>hello there</span></div>
<div class="test" id="testW"><span>缔造真正全球通行       	   的万维网</span></div>
<div class="ref" id="refW"><span>缔造真正全球通行 的万维网</span></div>
<div class="test" id="testF"><span>ＦＵＬＬ   	        ＷＩＤＴＨ</span></div>
<div class="ref" id="refF"><span>ＦＵＬＬ  ＷＩＤＴＨ</span></div>
<div class="test" id="testH"><span>ﾊﾝ   	         ｶｸ</span></div>
<div class="ref" id="refH"><span>ﾊﾝ   ｶｸ</span></div>
<script>
test(function() {
assert_equals(document.getElementById('test').firstChild.offsetWidth, document.getElementById('ref').firstChild.offsetWidth);
}, "spaces removed");
test(function() {
assert_equals(document.getElementById('testW').firstChild.offsetWidth, document.getElementById('refW').firstChild.offsetWidth);
}, "spaces removed between Wide characters");
test(function() {
assert_equals(document.getElementById('testF').firstChild.offsetWidth, document.getElementById('refF').firstChild.offsetWidth);
}, "spaces removed between Fullwidth characters");
test(function() {
assert_equals(document.getElementById('testH').firstChild.offsetWidth, document.getElementById('refH').firstChild.offsetWidth);
}, "spaces removed between Halfwidth characters");
</script>
<!-- Notes:
The assertion will fail if space is produced for any line in the test paragraph.
-->




</body></html>
